+++ scripts/helm/helmLint-svc-single.sh
@@ -12,7 +12,7 @@ help()
         [ -d | --debug ] Enable Helm template debug
         [ -m | --microservice ] Microservice defined in microservices folder
         [ -i | --image ] File with microservice image tag and digest
-        [ -o | --output ] Default output to predefined dir. Otherwise set to "console" to print linting output on terminal
+        [ -o | --output ] Default output to predefined dir. Otherwise set to "console" to print template output on terminal or set to a file path to redirect output
         [ -c | --clean ] Clean files and directories after script successfull execution
         [ -sd | --skip-dep ] Skip Helm dependencies setup
         [ -cp | --chart-path ] Path to Chart.yaml file (overrides environment selection; must be an existing file)
@@ -65,7 +65,7 @@ do
         -o | --output)
           [[ "${2:-}" ]] || "When specified, output cannot be null" || help
           output_redirect=$2
-          if [[ $output_redirect != "console" ]]; then
+          if [[ $output_redirect != "console" ]] && [[ -z "$output_redirect" ]]; then
             help
           fi
           step=2
@@ -127,7 +127,7 @@ fi
 ENV=$environment
 MICROSERVICE_DIR=$( echo $microservice | sed  's/-/_/g' )
 OUT_DIR="$ROOT_DIR/out/lint/$ENV/microservice_$MICROSERVICE_DIR"
-if [[ $output_redirect != "console" ]]; then
+if [[ "$output_redirect" != "console" ]] && [[ -z "$output_redirect" ]]; then
   rm -rf "$OUT_DIR"
   mkdir  -p "$OUT_DIR"
 else
@@ -150,6 +150,8 @@ fi
 OUTPUT_TO="> \"$OUT_DIR/$microservice.out.yaml\""
 if [[ $output_redirect == "console" ]]; then
   OUTPUT_TO=""
+elif [[ -n "$output_redirect" ]]; then
+  OUTPUT_TO="> \"$output_redirect\""
 fi
 #LINT_CMD=$LINT_CMD" \"$ROOT_DIR/charts/interop-eks-microservice-chart\" -f \"$ROOT_DIR/commons/$ENV/values-microservice.compiled.yaml\" -f \"$ROOT_DIR/microservices/$microservice/$ENV/values.yaml\" --set enableLookup=false $OUTPUT_TO"
 LINT_CMD+=" \"$ROOT_DIR/charts/interop-eks-microservice-chart\""
@@ -160,6 +162,6 @@ LINT_CMD+=" $OUTPUT_TO"
 
 eval $LINT_CMD
 
-if [[ $output_redirect != "console" && $post_clean == true ]]; then
+if [[ $output_redirect != "console" ]] && [[ -z "$output_redirect" ]] && [[ $post_clean == true ]]; then
   rm -rf $OUT_DIR
 fi
