+++ scripts/helm/helmTemplate-svc-single.sh
@@ -12,7 +12,7 @@ help()
         [ -d | --debug ] Enable Helm template debug
         [ -m | --microservice ] Microservice defined in microservices folder
         [ -i | --image ] File with microservice image tag and digest
-        [ -o | --output ] Default output to predefined dir. Otherwise set to "console" to print template output on terminal
+        [ -o | --output ] Default output to predefined dir. Otherwise set to "console" to print template output on terminal or set to a file path to redirect output
         [ -c | --clean ] Clean files and directories after script successfull execution
         [ -v | --verbose ] Show debug messages
         [ -sd | --skip-dep ] Skip Helm dependencies setup
@@ -69,7 +69,7 @@ do
         -o | --output)
           [[ "${2:-}" ]] || "When specified, output cannot be null" || help
           output_redirect=$2
-          if [[ $output_redirect != "console" ]]; then
+          if [[ $output_redirect != "console" ]] && [[ -z "$output_redirect" ]]; then
             help
           fi
           step=2
@@ -141,7 +141,7 @@ ENV=$environment
 MICROSERVICE_DIR=$( echo $microservice | sed  's/-/_/g' )
 
 OUT_DIR="$ROOT_DIR/out/templates/$ENV/microservice_$MICROSERVICE_DIR"
-if [[ $output_redirect != "console" ]]; then
+if [[ "$output_redirect" != "console" ]] && [[ -z "$output_redirect" ]]; then
   rm -rf "$OUT_DIR"
   mkdir  -p "$OUT_DIR"
 else
@@ -171,6 +171,8 @@ OUTPUT_FILE="\"$OUT_DIR/$microservice.out.yaml\""
 OUTPUT_TO="> $OUTPUT_FILE"
 if [[ $output_redirect == "console" ]]; then
   OUTPUT_TO=""
+elif [[ -n "$output_redirect" ]]; then
+  OUTPUT_TO="> \"$output_redirect\""
 fi
 
 #TEMPLATE_CMD=$TEMPLATE_CMD" $microservice interop-eks-microservice-chart/interop-eks-microservice-chart -f \"$ROOT_DIR/commons/$ENV/values-microservice.compiled.yaml\" -f \"$ROOT_DIR/microservices/$microservice/$ENV/values.yaml\" $OUTPUT_TO"
@@ -181,6 +183,6 @@ if [[ $verbose == true ]]; then
   echo "Successfully created Helm Template for microservice $microservice at $OUTPUT_FILE"
 fi
 
-if [[ $output_redirect != "console" && $post_clean == true ]]; then
+if [[ $output_redirect != "console" ]] && [[ -z "$output_redirect" ]] && [[ $post_clean == true ]]; then
   rm -rf $OUT_DIR
 fi
